Resume Parsering/requirements.txt
pandas
python-docx
PyMuPDF
spacy
phonenumbers
email_validator
openpyxl
